[MIG] pos_supplierinfo_search: Migration to 17.0
