+++ pos_supplierinfo_search/README.rst
@@ -70,6 +70,10 @@ Contributors
 
   - Carlos Lopez
 
+- `Heliconia Solutions Pvt. Ltd. <https://www.heliconia.io>`__
+
+  - Bhavesh Heliconia
+
 Maintainers
 -----------
