+++ pos_supplierinfo_search/__manifest__.py
@@ -3,7 +3,7 @@
 {
     "name": "Point of sale - Search products by supplier",
     "summary": "Search products by supplier data",
-    "version": "16.0.1.0.0",
+    "version": "17.0.1.0.0",
     "development_status": "Beta",
     "category": "Point Of Sale",
     "website": "https://github.com/OCA/pos",
@@ -17,8 +17,8 @@
         "point_of_sale",
     ],
     "assets": {
-        "point_of_sale.assets": [
-            "/pos_supplierinfo_search/static/src/js/db.js",
+        "point_of_sale._assets_pos": [
+            "/pos_supplierinfo_search/static/src/js/db.esm.js",
         ],
     },
 }
