+++ pos_supplierinfo_search/readme/CONTRIBUTORS.md
@@ -2,3 +2,5 @@
 - Dhara Solanki \<<dhara.solanki@initos.com>\>
 - \[Tecnativa\](<https://www.tecnativa.com>):
   - Carlos Lopez
+- [Heliconia Solutions Pvt. Ltd.](https://www.heliconia.io)
+  - Bhavesh Heliconia
