+++ pos_supplierinfo_search/static/description/index.html
@@ -416,6 +416,10 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <li>Carlos Lopez</li>
 </ul>
 </li>
+<li><a class="reference external" href="https://www.heliconia.io">Heliconia Solutions Pvt. Ltd.</a><ul>
+<li>Bhavesh Heliconia</li>
+</ul>
+</li>
 </ul>
 </div>
 <div class="section" id="maintainers">
